CYB-12: Implement fetching product by id and related products by category

+++ src/app/api/products/[id]/route.ts
@@ -1,7 +1,40 @@
 import { NextRequest, NextResponse } from 'next/server';
 
 import { updateProductSchema } from '@/entities/product';
-import { deleteProduct, updateProduct } from '@/entities/product/server';
+import {
+  deleteProduct,
+  getProductById,
+  updateProduct,
+} from '@/entities/product/server';
+import { uuidSchema } from '@/shared/model';
+
+export async function GET(
+  request: NextRequest,
+  { params }: { params: Promise<{ id: string }> },
+) {
+  try {
+    const id = (await params).id;
+    if (!uuidSchema.safeParse(id).success) {
+      return NextResponse.json(
+        { error: 'Invalid product ID format' },
+        { status: 400 },
+      );
+    }
+    const product = await getProductById(id);
+
+    if (!product) {
+      return NextResponse.json({ error: 'Product not found' }, { status: 404 });
+    }
+
+    return NextResponse.json(product);
+  } catch (error) {
+    console.error('Error fetching product by ID:', error);
+    return NextResponse.json(
+      { error: 'Internal Server Error' },
+      { status: 500 },
+    );
+  }
+}
 
 export async function PUT(
   request: NextRequest,
@@ -9,17 +42,19 @@ export async function PUT(
 ) {
   try {
     const id = (await params).id;
+    if (!uuidSchema.safeParse(id).success) {
+      return NextResponse.json(
+        { error: 'Invalid product ID format' },
+        { status: 400 },
+      );
+    }
     const body = await request.json();
     const parsedData = updateProductSchema.parse(body);
     const updatedProduct = await updateProduct(id, parsedData);
     return NextResponse.json(updatedProduct);
   } catch (error) {
-    return NextResponse.json(
-      {
-        error: error instanceof Error ? error.message : 'Internal Server Error',
-      },
-      { status: 400 },
-    );
+    console.error('Error updating product:', error);
+    return NextResponse.json({ error: 'Bad Request' }, { status: 400 });
   }
 }
 
@@ -29,13 +64,18 @@ export async function DELETE(
 ) {
   try {
     const id = (await params).id;
+    if (!uuidSchema.safeParse(id).success) {
+      return NextResponse.json(
+        { error: 'Invalid product ID format' },
+        { status: 400 },
+      );
+    }
     await deleteProduct(id);
     return new NextResponse(null, { status: 204 });
   } catch (error) {
+    console.error('Error deleting product:', error);
     return NextResponse.json(
-      {
-        error: error instanceof Error ? error.message : 'Internal Server Error',
-      },
+      { error: 'Internal Server Error' },
       { status: 500 },
     );
   }
